fix: change photo type and split editor modal on components
