+++ features/posts/PostPhotoSelectModal.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import React, {useState} from "react";
 import styled from "styled-components";
 import {baseTheme} from "styles/styledComponents/theme";
-import PostPhotoEditorModal from "./PostPhotoEditorModal";
+import PostPhotoEditorModal, { PhotoType } from "./PostPhotoEditorModal";
 import {Button} from "common/components/Button/Button";
 import {ThemeButton} from "common/enums/themeButton";
 
@@ -19,13 +19,13 @@ const PostPhotoSelectModal = ({
   handleFullScreen: (full: boolean) => void;
   isHeaderOpen?: boolean
   isHeader2Open?: boolean
-  photoPostProps?: string[]
+  photoPostProps?: PhotoType[]
 }) => {
   const [photo, setPhoto] = useState<File>(); // изображение, передаваемое в компоненту редактирования
   const [isEditorOpen, setIsEditorOpen] = useState(isHeaderOpen); // открытие модального окна для редактирования
-  const [isFilterOpen, setIsFilterOpen] = useState(false) // открытие модального окна для наложения фильтров
-  const [photoList, setPhotoList] = useState([''])
-
+  // const [isFilterOpen, setIsFilterOpen] = useState(false) // открытие модального окна для наложения фильтров
+  // const [photoList, setPhotoList] = useState([''])
+  
   const handleSelectPhoto = (e: React.ChangeEvent<HTMLInputElement>) => {
     if (e.target.files?.length) {
       const file = e.target.files[0];
@@ -45,11 +45,6 @@ const PostPhotoSelectModal = ({
     handleFullScreen(full);
   };
 
-  const handleFilterOpen = (editPhotoList: string[]) => {
-    setIsFilterOpen(true)
-    setPhotoList(editPhotoList)
-  }
-
   return (
     <>
       <StyledModalBody>
@@ -59,7 +54,6 @@ const PostPhotoSelectModal = ({
             handleEditorClose={handleEditorClose}
             handleFullScreen={(full) => handleClickFullScreen(full)}
             photoPost1={photoPostProps || []}
-            handleFilterOpen={handleFilterOpen}
           />
           : (
             <>
